backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
neo4j==5.27.0
pydantic==2.10.5
pydantic-settings==2.7.1
langgraph==0.2.60
google-generativeai==0.8.4
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
mcp[cli]==1.3.0
